# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /billing/stats:
    get:
      operationId: getBillingStats
      tags: [billing]
      summary: Billing summary stats by status and deployment type
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Billing statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingStats"
  /billing/documents:
    get:
      operationId: listBillingDocuments
      tags: [billing]
      summary: List invoices and quotations
      parameters:
        - name: kind
          in: query
          schema:
            type: string
            enum: [invoice, quotation]
        - name: search
          in: query
          schema:
            type: string
        - name: deploymentTypeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of billing documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillingDocumentSummary"
    post:
      operationId: createBillingDocument
      tags: [billing]
      summary: Create an invoice or quotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingDocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingDocumentCreated"
        "400":
          description: Invalid input
  /billing/documents/{id}:
    get:
      operationId: getBillingDocument
      tags: [billing]
      summary: Get an invoice or quotation with line items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingDocument"
        "404":
          description: Not found
    patch:
      operationId: updateBillingDocument
      tags: [billing]
      summary: Update an invoice or quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingDocumentUpdate"
      responses:
        "204":
          description: Updated
        "400":
          description: Invalid input
        "404":
          description: Not found
    delete:
      operationId: deleteBillingDocument
      tags: [billing]
      summary: Delete an invoice or quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BillingStats:
      type: object
      required: [totalInvoices, totalQuotations, draft, sent, paid, void, totalRevenue]
      properties:
        totalInvoices:
          type: integer
        totalQuotations:
          type: integer
        draft:
          type: integer
        sent:
          type: integer
        paid:
          type: integer
        void:
          type: integer
        totalRevenue:
          type: string
        byDeploymentType:
          type: array
          items:
            type: object
            properties:
              deploymentTypeId:
                type: ["integer", "null"]
              deploymentTypeName:
                type: ["string", "null"]
              count:
                type: integer
              revenue:
                type: string
...
    BillingDocumentSummary:
      type: object
      required: [id, kind, number, companyId, companyName, customerName, issueDate, gstRate, gstInclusive, status, subtotal, createdAt, updatedAt]
      properties:
        id: { type: integer }
        kind: { type: string, enum: [invoice, quotation] }
        number: { type: string }
        companyId: { type: integer }
        companyName: { type: string }
        customerName: { type: string }
        customerAddress: { type: ["string", "null"] }
        customerTin: { type: ["string", "null"] }
        issueDate: { type: string }
        dueDate: { type: ["string", "null"] }
        terms: { type: ["string", "null"] }
        gstRate: { type: string }
        gstInclusive: { type: boolean }
        notes: { type: ["string", "null"] }
        status: { type: string }
        deploymentTypeId: { type: ["integer", "null"] }
        subtotal: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    BillingItem:
      type: object
      required: [id, documentId, position, description, qty, rate, amount]
      properties:
        id: { type: integer }
        documentId: { type: integer }
        position: { type: integer }
        description: { type: string }
        detail: { type: ["string", "null"] }
        qty: { type: string }
        rate: { type: string }
        amount: { type: string }
    BillingDocument:
      allOf:
        - $ref: "#/components/schemas/BillingDocumentSummary"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/BillingItem"
    BillingItemInput:
      type: object
      required: [description]
      properties:
        description: { type: string, minLength: 1 }
        detail: { type: string }
        qty: { type: string }
        rate: { type: string }
    BillingDocumentInput:
      type: object
      required: [kind, companyId, customerName, issueDate, items]
      properties:
        kind: { type: string, enum: [invoice, quotation] }
        companyId: { type: integer }
        customerName: { type: string, minLength: 1 }
        customerAddress: { type: string }
        customerTin: { type: string }
        issueDate: { type: string }
        dueDate: { type: string }
        terms: { type: string }
        gstRate: { type: string }
        gstInclusive: { type: boolean }
        notes: { type: string }
        status: { type: string }
        deploymentTypeId: { type: ["integer", "null"] }
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/BillingItemInput"
    BillingDocumentUpdate:
      type: object
      properties:
        companyId: { type: integer }
        customerName: { type: string, minLength: 1 }
        customerAddress: { type: ["string", "null"] }
        customerTin: { type: ["string", "null"] }
        issueDate: { type: string }
        dueDate: { type: ["string", "null"] }
        terms: { type: ["string", "null"] }
        gstRate: { type: string }
        gstInclusive: { type: boolean }
        notes: { type: ["string", "null"] }
        status: { type: string }
        deploymentTypeId: { type: ["integer", "null"] }
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/BillingItemInput"
    BillingDocumentCreated:
      type: object
      required: [id]
      properties:
        id: { type: integer }